URLに異常な数値が入力された際に例外ぺージを表示させるように修正

+++ src/main/java/jp/co/rakus/ecommerce_c/service/DetailService.java
@@ -39,7 +39,7 @@ public class DetailService {
 	 * @param id 商品ID
 	 * @return item 検索した商品データ
 	 */
-	public Item loadItem(Integer id) {
+	public Item loadItem(long id) {
 		return itemRepository.loadItem(id);
 	}
